pyproject.toml
[tool.poetry]
name = "librarian"
version = "0.1.0"
description = ""
authors = ["kk6 <hiro.ashiya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tortoise-orm = "^0.11.8"
tqdm = "^4.31"
aiohttp = "^3.5"
bocadillo = "^0.14.1"
bocadillo-cli = "^0.2.1"

[tool.poetry.dev-dependencies]
pytest = "^3.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
